[fix] 사용자의 반려동물이 없을 경우 대표반려동물 설정, 소수점 설정 에러 부분 수정

+++ src/main/java/com/kkosoonnae/jpa/projection/StoreDetailViewProjection.java
@@ -21,7 +21,7 @@ public record StoreDetailViewProjection(
 ) {
     public StoreDetailViewProjection(Integer storeNo, String storeName, String content, String phone,
                                      String roadAddress, LocalTime openingTime, LocalTime closingTime,
-                                     Double averageScope, Long totalLikeStore) {
+                                     Integer averageScope, Long totalLikeStore) {
         this(storeNo, storeName, content, phone, roadAddress, openingTime, closingTime,
                 new ArrayList<>(), averageScope, totalLikeStore);
     }
